requirements.txt
Flask
python-dotenv
google-generativeai
fuzzywuzzy 